dfjdjf

+++ TransportClasses.cpp
@@ -13,8 +13,7 @@ Transport::Transport(string b, int s) {
   speed = s;
 }
 
-Transport::~Transport() {
-}
+Transport::~Transport() {}
 
 int Transport::SetBrand(string b) {
   brand = b;
@@ -26,33 +25,28 @@ int Transport::SetSpeed(int s) {
   return 1;
 }
 
-int Transport::SetSpeed(int minSpeed, int maxSpeed) {
+short int Transport::SetSpeed(short int minSpeed, short int maxSpeed) {
   speed = (minSpeed + maxSpeed) / 2;
   return 1;
 }
 
 void Transport::ShowInfo() const {
-  cout << "Транспорт марки " << brand << ", швидкість " << speed << " км/год.\n";
+  cout << "Транспорт марки " << brand << ", швидкість " << speed
+       << " км/год.\n";
 }
 
-void Transport::Move() const {
-  cout << "Транспорт рухається.\n";
-}
+void Transport::Move() const { cout << "Транспорт рухається.\n"; }
 
-Car::Car(string b, int s) : Transport(b, s) {
-}
+Car::Car(string b, int s) : Transport(b, s) {}
 
 void Car::ShowInfo() const {
-  cout << "Автомобіль марки " << brand << ", швидкість " << speed << " км/год.\n";
+  cout << "Автомобіль марки " << brand << ", швидкість " << speed
+       << " км/год.\n";
 }
 
-void Car::Move() const {
-  cout << "Автомобіль їде по дорозі.\n";
-}
+void Car::Move() const { cout << "Автомобіль їде по дорозі.\n"; }
 
-Bus::Bus(string b, int s, int p) : Transport(b, s) {
-  SetPassengers(p);
-}
+Bus::Bus(string b, int s, int p) : Transport(b, s) { SetPassengers(p); }
 
 int Bus::SetPassengers(int p) {
   passengers = p;
@@ -64,9 +58,7 @@ void Bus::ShowInfo() const {
        << " км/год, пасажирів " << passengers << ".\n";
 }
 
-void Bus::Move() const {
-  cout << "Автобус зупиняється на зупинках.\n";
-}
+void Bus::Move() const { cout << "Автобус зупиняється на зупинках.\n"; }
 
 Truck::Truck(string b, int s, int load) : Transport(b, s) {
   SetLoadCapacity(load);
@@ -82,9 +74,7 @@ void Truck::ShowInfo() const {
        << " км/год, вантажність " << loadCapacity << " тонн.\n";
 }
 
-void Truck::Move() const {
-  cout << "Вантажівка перевозить вантаж.\n";
-}
+void Truck::Move() const { cout << "Вантажівка перевозить вантаж.\n"; }
 
 ElectricCar::ElectricCar(string b, int batteryPower) : Transport(b, 0) {
   SetBattery(batteryPower);
@@ -96,8 +86,8 @@ int ElectricCar::SetBattery(int batteryPower) {
 }
 
 void ElectricCar::ShowInfo() const {
-  cout << "Електромобіль марки " << brand << ", батарея "
-       << battery << " кВт*год, швидкість " << speed << " км/год.\n";
+  cout << "Електромобіль марки " << brand << ", батарея " << battery
+       << " кВт*год, швидкість " << speed << " км/год.\n";
 }
 
 void ElectricCar::Move() const {
